fix(message): 修复消息相关DTO字段类型
- 将QueryLatestMessagesRequest中的botId和targetId字段从Integer改为Long类型
- 将QueryMessagesByMessageIdRequest中的botId和targetId字段从Integer改为Long类型
- 将QueryMessagesBySenderRequest中的botId和senderId字段从Integer改为Long类型
- 将SearchMessagesRequest中的botId和targetId字段从Integer改为Long类型
- 将UpdateMessageRequest中的id字段从Integer改为Long类型

+++ MangoBot-cli/src/main/java/io/github/mangomaner/mangobot/model/dto/message/QueryLatestMessagesRequest.java
@@ -10,9 +10,9 @@ public class QueryLatestMessagesRequest {
 
     @Schema(description = "Bot ID")
     @NotNull(message = "Bot ID不能为空")
-    private Integer botId;
+    private Long botId;
 
     @Schema(description = "群组ID或好友ID")
     @NotNull(message = "群组ID或好友ID不能为空")
-    private Integer targetId;
+    private Long targetId;
 }
